Module Atable

type t = string Agrid.t

The type of tables

module type Config = sig ... end

The signature of modules expected by the Make functor

module type S = sig ... end

The signature of modules produced by the Make functor

module Make : functor (Config : Config) -> S

The Make functor