Parameter Make.1-Config

val initial : t Stdlib.Option.t

The initial table

val table_id : string

The HTML id of the table

val add_row_button_id : string Stdlib.Option.t

The HTML id of the button to add a row

val add_col_button_id : string Stdlib.Option.t

The HTML id of the button to add a column

val rm_row_button_id : string Stdlib.Option.t

The HTML id of the button to remove a row

val rm_col_button_id : string Stdlib.Option.t

The HTML id of the button to remove a column

val reset_button_id : string Stdlib.Option.t

The HTML id of the button to reset the table to its initial value

val action_button_id : string Stdlib.Option.t

The HTML id of the button to run a user-defined action

val action_fun : (t -> unit) Stdlib.Option.t

The user-defined action which will be run with the action button