Module type Atable.S

The signature of modules produced by the Make functor

val init : unit -> unit

The function to initialize the table, must be called once the page is loaded

val get_content : unit -> t

Return the current content of the table

val dim : unit -> int * int

Return the current dimensions of the table

val add_col : unit -> unit

Add an empty column to the table

val add_row : unit -> unit

Add an empty row to the table

val rm_col : unit -> unit

Remove a column to the table

val rm_row : unit -> unit

Remove a row to the table

val reset : unit -> unit

Reset the table to its initial value